update offsetInMilliseconds to type long

+++ Alexa.NET/Response/Directive/AudioItemStream.cs
@@ -17,6 +17,6 @@ namespace Alexa.NET.Response.Directive
 
         [JsonRequired]
         [JsonProperty("offsetInMilliseconds")]
-        public int OffsetInMilliseconds { get; set; }
+        public long OffsetInMilliseconds { get; set; }
     }
 }
